Aiohttp/requirements.txt
aiohttp==3.9.3
aiohttp_cors==0.7.0
sqlalchemy[asyncio]==2.0.25
aiosqlite==0.19.0
pydantic==2.5.3
python-dotenv==1.0.0
bcrypt==4.1.2
pyjwt==2.8.0